requirement.txt
streamlit
watchdog
openai
requests
azure-ai-formrecognizer
fastapi
uvicorn
python-dotenv
databases
asyncpg
beautifulsoup4
sqlalchemy>=2.0
psycopg2-binary>=2.9